fix(inspector): thread childIndex and innerCards into ed25519 and zk-elgamal cards
Widen SiblingInstructionData to ReadonlyUint8Array so the base58 encoder result
needs no cast.

+++ app/features/decode-instruction-ed25519/lib/sibling-data.ts
@@ -14,7 +14,7 @@ export function siblingDataFromParsedTransaction(tx: ParsedTransaction): Sibling
             return undefined;
         }
         try {
-            return BASE58_ENCODER.encode(target.data) as Uint8Array;
+            return BASE58_ENCODER.encode(target.data);
         } catch {
             return undefined;
         }
